+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # 0.11.0
 
-* Unnecessary and deprecated features removed
-
+* Adds `app` with `HullApp`, `Server` and `Worker`
+* Adds `helpers` with functions being added to the `req.hull.client` after being initiated by the `Hull.Middleware`
+* Adds `infra` with `Instrumentation`, `Queue`, `Cache` and `Batcher` services
+* Adds `utils` with `handlers`, `middlewares` and `routers` used as a building blocks of the `HullApp` and to be used by the ship/connector directly
 
 
 # 0.10.7